requirements.txt
flask==2.0.1
flask-cors==3.0.10
selenium==4.1.0
webdriver-manager==3.5.2
Pillow==8.3.2
opencv-python==4.5.3.56
numpy==1.21.2 